% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg_topoplots.R
\name{maketopoplot}
\alias{maketopoplot}
\title{A topoplot Function}
\usage{
maketopoplot(
  dataset = dat,
  condition = NULL,
  conditionlevels = levels(as.factor(dataset[, condition])),
  differenceplots = T,
  differencelevel = "all",
  extralegend = T,
  addlabels = F,
  method = "akima",
  simsignals = T,
  quick = T,
  palette = "RdBu",
  signallimits = NULL,
  signalcomplimits = NULL,
  markchans = NULL,
  nrchans = 64,
  chanmarkcol = "black",
  plottitle = "",
  marksize = 1.1,
  textsize = 10,
  textfamily = NULL
)
}
\arguments{
\item{dataset}{the data set to plot}

\item{condition}{name of the variable that specifies the condition to plot}

\item{conditionlevels}{if not specified, all levels available from the unique values of the condition will be used. (currently max 4)}

\item{differenceplots}{calculate differences between condition levels and plot them. Defaults to TRUE.}

\item{differencelevel}{specify a specific condition as the reference level. Defaults to "all" for all pairwise comparisons.}

\item{extralegend}{plot legend extra (vs. under the label) to the figure. Defaults to TRUE.}

\item{addlabels}{include the labels of the electrode channels. Defaults to FALSE.}

\item{method}{one of "akima" or "biharmonic". Akima performs faster and includes a spline method but requires a specific version of the akima package (not newer than 0.6-2.1). Defaults to "akima".}

\item{simsignals}{Defaults to TRUE.}

\item{quick}{faster plot with lower resolution. Defaults to TRUE.}

\item{palette}{colour palette of the topoplot (e.g., BrBG, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral). Defaults to "RdBu".}

\item{signallimits}{lmit the colour palette signals for all plots to make them comparable. Defaults to NULL.}

\item{signalcomplimits}{same as signallimits but for the difference plots. Defaults to NULL.}

\item{markchans}{a character vector specifying the names of the to be highlighted ROI channels. Defaults to NULL.}

\item{nrchans}{choose 64 or 128. 64 includes the extended 10-20 system layour, 128 the ABCD layout. Defaults to 64.}

\item{chanmarkcol}{colour of the ROIchannels if to highlight. Defaults to "black".}

\item{plottitle}{title of the plot, Defaults to ''}

\item{marksize}{size of the electrode markers. Defaults to 1.1}

\item{textsize}{size of the text capture}

\item{textfamily}{if specified, the package "extrafont" needs to be installed and loaded.}
}
\description{
This function allows you to plot ERP topoplots for different conditions, specify the printed limits (microvolt scale) and customize the appearance (colours, fonts, etc.).
}
\examples{




data(xdat) #load example data frame
maketopoplot(dataset = xdat, condition = "difficulty", nrchans = 128)
}
\keyword{topo}
